Move adddata and getdata functions to module
