+++ src/js/index.js
@@ -1,42 +1,15 @@
 /** @format */
-
-// import '../scss/style.scss';
-// import displayBoard from './displayBoard.js';
-
-// ! Add data to api make it module
-const APIURL = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api';
-const gamesID = 'nodsadı'; // YUELB6qMFZSbsC3Ub4pp // Super Mario
-const scoreUrl = `${APIURL}/games/${gamesID}/scores`;
-
-const addDataApı = (user, score) => {
-  fetch(scoreUrl, {
-    method: 'POST',
-    body: JSON.stringify({
-      user,
-      score,
-    }),
-    headers: {
-      'Content-type': 'application/json; charset=UTF-8',
-    },
-  });
-};
+import '../scss/style.scss';
+import { getApiData, addDataApi } from './leaderBoardApi.js';
 
 const addScore = (name, score) => {
   name = document.getElementById('name').value;
   score = document.getElementById('score').value;
-  addDataApı(name, score);
+  addDataApi(name, score);
 };
 
 const submitBtn = document.getElementById('submit-btn');
 submitBtn.addEventListener('click', addScore);
-
-// ! GET DATA FROM API
-
-const getApiData = async () => {
-  const response = await fetch(scoreUrl);
-  const data = await response.json();
-  return data.result;
-};
 
 const leaderBoard = document.getElementsByClassName('leader-board');
